# docs/api/openapi.generated.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/invitations/{token}/accept": {
      "post": {
        "operationId": "acceptInvitation",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 32
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ACTIVE MEMBER membership created or reactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "oidcBearer": []
          }
        ],
        "summary": "Accept a single-use invitation",
        "tags": [
          "Households"
        ]
      }
    }
  },
  "info": {
    "title": "TurnTable API",
    "description": "TurnTable API-first modular monolith implementation contract",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "oidcBearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "OIDC access token"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "display_name": {
            "type": "string",
            "example": "Alex Rivera"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          }
        },
        "required": [
          "id",
          "display_name"
        ]
      },
...
      "ProblemErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://api.turntable.example/problems/validation-error"
          },
          "title": {
            "type": "string",
            "example": "Request validation failed"
          },
          "status": {
            "type": "number",
            "example": 422
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "detail": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemErrorDto"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "trace_id"
        ]
      },
...
      "MemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "household_id": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MEMBER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "LEFT",
              "REMOVED"
            ]
          },
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "left_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "id",
          "household_id",
          "user",
          "role",
          "status",
          "joined_at",
          "left_at",
          "version"
        ]
      },
...
    }
  }
}